+++ ext/include/linear_algebra.c
@@ -5,6 +5,7 @@
 #include <php.h>
 #include <cblas.h>
 #include <lapacke.h>
+#include <stdlib.h>
 
 #include "php_ext.h"
 #include "kernel/operators.h"
@@ -108,6 +109,7 @@ void tensor_inverse(zval * return_value, zval * a)
     unsigned int n = zend_array_count(aHat);
 
     double * va = emalloc(n * n * sizeof(double));
+    int * pivots = emalloc(n * sizeof(int));
 
     for (i = 0; i < n; ++i) {
         row = Z_ARR(ba[i].val)->arData;
@@ -118,8 +120,6 @@ void tensor_inverse(zval * return_value, zval * a)
     }
 
     lapack_int status;
-
-    int pivots[n];
 
     status = LAPACKE_dgetrf(LAPACK_ROW_MAJOR, n, n, va, n, pivots);
 
@@ -140,6 +140,127 @@ void tensor_inverse(zval * return_value, zval * a)
 
         for (j = 0; j < n; ++j) {
             add_next_index_double(&rowB, va[i * n + j]);
+        }
+
+        add_next_index_zval(&b, &rowB);
+    }
+
+    RETVAL_ARR(Z_ARR(b));
+
+    efree(va);
+}
+
+void tensor_ref(zval * return_value, zval * a)
+{
+    unsigned int i, j;
+    Bucket * row;
+    zval rowB, b;
+    zval tuple;
+
+    zend_zephir_globals_def * zephir_globals = ZEPHIR_VGLOBAL;
+
+    openblas_set_num_threads(zephir_globals->num_threads);
+
+    zend_array * aHat = Z_ARR_P(a);
+
+    Bucket * ba = aHat->arData;
+
+    unsigned int m = zend_array_count(aHat);
+    unsigned int n = zend_array_count(Z_ARR(ba[0].val));
+
+    double * va = emalloc(m * n * sizeof(double));
+    int * pivots = emalloc(MIN(m, n) * sizeof(int));
+
+    for (i = 0; i < m; ++i) {
+        row = Z_ARR(ba[i].val)->arData;
+
+        for (j = 0; j < n; ++j) {
+            va[i * n + j] = zephir_get_doubleval(&row[j].val);
+        }
+    }
+
+    lapack_int status = LAPACKE_dgetrf(LAPACK_ROW_MAJOR, m, n, va, n, pivots);
+
+    if (status != 0) {
+        RETURN_NULL();
+    }
+    
+    array_init_size(&b, m);
+
+    long swaps = 0;
+
+    for (i = 0; i < m; ++i) {
+        array_init_size(&rowB, n);
+
+        for (j = 0; j < i; ++j) {
+            add_next_index_double(&rowB, 0.0);
+        }
+
+        for (j = i; j < n; ++j) {
+            add_next_index_double(&rowB, va[i * n + j]);
+        }
+
+        add_next_index_zval(&b, &rowB);
+
+        if (i + 1 != pivots[i]) {
+            ++swaps;
+        }
+    }
+
+    array_init_size(&tuple, 2);
+    
+    add_next_index_zval(&tuple, &b);
+    add_next_index_long(&tuple, swaps);
+
+    RETVAL_ARR(Z_ARR(tuple));
+
+    efree(va);
+    efree(pivots);
+}
+
+void tensor_cholesky(zval * return_value, zval * a)
+{
+    unsigned int i, j;
+    Bucket * row;
+    zval rowB, b;
+
+    zend_zephir_globals_def * zephir_globals = ZEPHIR_VGLOBAL;
+
+    openblas_set_num_threads(zephir_globals->num_threads);
+
+    zend_array * aHat = Z_ARR_P(a);
+
+    Bucket * ba = aHat->arData;
+
+    unsigned int n = zend_array_count(aHat);
+
+    double * va = emalloc(n * n * sizeof(double));
+
+    for (i = 0; i < n; ++i) {
+        row = Z_ARR(ba[i].val)->arData;
+
+        for (j = 0; j < n; ++j) {
+            va[i * n + j] = zephir_get_doubleval(&row[j].val);
+        }
+    }
+
+    lapack_int status = LAPACKE_dpotrf(LAPACK_ROW_MAJOR, 'L', n, va, n);
+
+    if (status != 0) {
+        RETURN_NULL();
+    }
+    
+    array_init_size(&b, n);
+
+    for (i = 0; i < n; ++i) {
+        array_init_size(&rowB, n);
+
+        for (j = 0; j <= i; ++j) {
+            add_next_index_double(&rowB, va[i * n + j]);
+        }
+
+        for (j = i + 1; j < n; ++j) {
+            add_next_index_double(&rowB, 0.0);
         }
 
         add_next_index_zval(&b, &rowB);