+++ ext/tensor/reductions/ref.zep.h
@@ -4,8 +4,6 @@ extern zend_class_entry *tensor_reductions_ref_ce;
 ZEPHIR_INIT_CLASS(Tensor_Reductions_Ref);
 
 PHP_METHOD(Tensor_Reductions_Ref, reduce);
-PHP_METHOD(Tensor_Reductions_Ref, gaussianElimination);
-PHP_METHOD(Tensor_Reductions_Ref, rowReductionMethod);
 PHP_METHOD(Tensor_Reductions_Ref, __construct);
 PHP_METHOD(Tensor_Reductions_Ref, a);
 PHP_METHOD(Tensor_Reductions_Ref, swaps);
@@ -14,22 +12,6 @@ PHP_METHOD(Tensor_Reductions_Ref, swaps);
 ZEND_BEGIN_ARG_WITH_RETURN_OBJ_INFO_EX(arginfo_tensor_reductions_ref_reduce, 0, 1, Tensor\\Reductions\\Ref, 0)
 #else
 ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_tensor_reductions_ref_reduce, 0, 1, IS_OBJECT, "Tensor\\Reductions\\Ref", 0)
-#endif
-	ZEND_ARG_OBJ_INFO(0, a, Tensor\\Matrix, 0)
-ZEND_END_ARG_INFO()
-
-#if PHP_VERSION_ID >= 70200
-ZEND_BEGIN_ARG_WITH_RETURN_OBJ_INFO_EX(arginfo_tensor_reductions_ref_gaussianelimination, 0, 1, Tensor\\Reductions\\Ref, 0)
-#else
-ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_tensor_reductions_ref_gaussianelimination, 0, 1, IS_OBJECT, "Tensor\\Reductions\\Ref", 0)
-#endif
-	ZEND_ARG_OBJ_INFO(0, a, Tensor\\Matrix, 0)
-ZEND_END_ARG_INFO()
-
-#if PHP_VERSION_ID >= 70200
-ZEND_BEGIN_ARG_WITH_RETURN_OBJ_INFO_EX(arginfo_tensor_reductions_ref_rowreductionmethod, 0, 1, Tensor\\Reductions\\Ref, 0)
-#else
-ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_tensor_reductions_ref_rowreductionmethod, 0, 1, IS_OBJECT, "Tensor\\Reductions\\Ref", 0)
 #endif
 	ZEND_ARG_OBJ_INFO(0, a, Tensor\\Matrix, 0)
 ZEND_END_ARG_INFO()
@@ -59,8 +41,6 @@ ZEND_END_ARG_INFO()
 
 ZEPHIR_INIT_FUNCS(tensor_reductions_ref_method_entry) {
 	PHP_ME(Tensor_Reductions_Ref, reduce, arginfo_tensor_reductions_ref_reduce, ZEND_ACC_PUBLIC|ZEND_ACC_STATIC)
-	PHP_ME(Tensor_Reductions_Ref, gaussianElimination, arginfo_tensor_reductions_ref_gaussianelimination, ZEND_ACC_PUBLIC|ZEND_ACC_STATIC)
-	PHP_ME(Tensor_Reductions_Ref, rowReductionMethod, arginfo_tensor_reductions_ref_rowreductionmethod, ZEND_ACC_PUBLIC|ZEND_ACC_STATIC)
 	PHP_ME(Tensor_Reductions_Ref, __construct, arginfo_tensor_reductions_ref___construct, ZEND_ACC_PUBLIC|ZEND_ACC_CTOR)
 	PHP_ME(Tensor_Reductions_Ref, a, arginfo_tensor_reductions_ref_a, ZEND_ACC_PUBLIC)
 	PHP_ME(Tensor_Reductions_Ref, swaps, arginfo_tensor_reductions_ref_swaps, ZEND_ACC_PUBLIC)